refactor: Parquet test
